Extend `responseDataPath` to support extracting objects inside arrays.
